third_party/blink/web_tests/external/wpt/editing/run/multitest_1001-2000-expected.txt

This is a testharness.js-based test.
Found 218 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["italic",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><i>a</i>bar</p>" but got "foo<div><span style=\\"font-style:italic\\">a</span>bar</div>"
[FAIL] [["italic",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<i>a</i>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["justifycenter",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifycenter",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifycenter",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<i>a</i>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["italic",""],["justifyfull",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyfull",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyfull",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<i>a</i>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["italic",""],["justifyleft",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyleft",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyleft",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["italic",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>bar" but got "fooabar"
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["italic",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["italic",""],["justifyright",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyright",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyright",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<i>a</i>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["italic",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["italic",""],["outdent",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["outdent",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>bar" but got "foo<span style=\\"font-style:italic\\">a</span>bar"
[FAIL] [["italic",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>bar" but got "foo<span style=\\"text-decoration-line:line-through\\">a</span>bar"
[FAIL] [["strikethrough",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["delete",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["delete",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["delete",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<s>a</s>bar" but got "foabar"
[FAIL] [["strikethrough",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<s>a</s>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["strikethrough",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["forwarddelete",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>ar" but got "fooaar"
[FAIL] [["strikethrough",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["strikethrough",""],["indent",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["indent",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["indent",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<s>a</s>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["strikethrough",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><s>a</s>bar" but got "foo<hr>abar"
[FAIL] [["strikethrough",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<s>a</s>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><s>a</s>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"text-decoration-line:line-through\\"><br></span>bar"
[FAIL] [["strikethrough",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><s>a</s>bar" but got "foo<span style=\\"text-decoration-line:line-through\\"><br>a</span>bar"
[FAIL] [["strikethrough",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<s>a</s>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["strikethrough",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><s>a</s>bar</p>" but got "foo<div><span style=\\"text-decoration-line:line-through\\">a</span>bar</div>"
[FAIL] [["strikethrough",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<s>a</s>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["justifycenter",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifycenter",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<s>a</s>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["strikethrough",""],["justifyfull",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyfull",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<s>a</s>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["strikethrough",""],["justifyleft",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyleft",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyleft",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["strikethrough",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>bar" but got "fooabar"
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["strikethrough",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["strikethrough",""],["justifyright",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyright",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<s>a</s>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["strikethrough",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["strikethrough",""],["outdent",""]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["outdent",""]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["strikethrough",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s>a</s>bar" but got "foo<span style=\\"text-decoration-line:line-through\\">a</span>bar"
[FAIL] [["strikethrough",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>bar" but got "foo<span style=\\"vertical-align:sub\\">a</span>bar"
[FAIL] [["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["delete",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["delete",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["delete",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<sub><font size=\\"3\\">a</font></sub>bar" but got "foabar"
[FAIL] [["subscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<sub>a</sub>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["subscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
